add tests for util func
