--- test/_old/multi-root-workspace-tests/MultiRootIndices.test.ts
@@ -1,6 +1,0 @@
-
-describe('Multi root indices', () => {
-
-    it('should be tested; will be done as soon as the feature is default enabled.');
-
-});
